fix: back to single-command CLI
