+++ src/practices/JavaScript/JsPackageManagementUsedPractice.ts
@@ -6,7 +6,7 @@ import { PracticeContext } from '../../contexts/practice/PracticeContext';
 @DxPractice({
   id: 'Javascript.PackageManagementUsed',
   name: 'Use JS Package Management',
-  impact: PracticeImpact.high,
+  impact: PracticeImpact.high, //which impact?
   suggestion: 'Use Package.json to keep track of packages that are being used in your application.',
   reportOnlyOnce: true,
   url: 'https://docs.npmjs.com/files/package.json',
