+++ src/scanner/ArgumentsProvider.ts
@@ -3,9 +3,9 @@ import { PracticeImpact } from '../model';
 export type ArgumentsProvider = {
   uri: string;
   auth: string | undefined;
-  json: boolean;
-  fail: PracticeImpact | 'all'; // optional as the AP is used also with command dxs practices
-  recursive: boolean; // optional as the AP is used also with command dxs practices
+  json: boolean | undefined;
+  fail: PracticeImpact | 'all';
+  recursive: boolean;
   ci: boolean;
   fix: boolean;
   fixPattern: string | undefined;
